more cleanup and comments
