+++ src/units/mod.rs
@@ -3,22 +3,25 @@
 mod activate;
 mod deactivate;
 mod dependency_resolving;
+mod from_parsed_config;
+mod id;
 mod insert_new;
 mod loading;
 mod remove;
-mod sanity_check;
-mod unit_parsing;
-mod units;
-mod from_parsed_config;
 mod runtime_info;
+mod sanity_check;
+mod status;
+mod unit;
+mod unit_parsing;
 
 pub use activate::*;
 pub use deactivate::*;
 pub use dependency_resolving::*;
+pub use id::*;
 pub use insert_new::*;
 pub use loading::load_all_units;
 pub use remove::*;
-pub use sanity_check::*;
-pub use unit_parsing::*;
-pub use units::*;
 pub use runtime_info::*;
+pub use sanity_check::*;
+pub use status::*;pub use unit::*;
+pub use unit_parsing::*;
